% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/accessfun.r
\name{list2db}
\alias{list2db}
\title{Save a List into a Database}
\usage{
list2db(channel, dat, ...)
}
\arguments{
\item{channel}{connection handle as returned by \code{\link{odbcConnect}}.}

\item{dat}{any valid list which is a list of data frames.}

\item{...}{arguments to be passed to \code{\link{sqlSave}}.}
}
\value{
Values are same with \code{\link{sqlSave}}.
}
\description{
Save a list into a *.mdb database. The list should be a combination of
data frames.
}

