% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_mdb}
\alias{read_mdb}
\title{Read MDB files}
\usage{
read_mdb(path, table_name = NULL, n_max = -1)
}
\arguments{
\item{path}{Path to the mdb/accdb file.}

\item{table_name}{Table to read. A string with the name of table to read. If
NULL, defaults to the first table.}

\item{n_max}{Maximum numbers of data rows to read. Use any negative number to read
all rows.}
}
\description{
Read MDB files
}
